changed code  from competition
