+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/commands/autonomous/BlueTop.java
@@ -6,6 +6,7 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.teamcode.systems.ArmSystem;
 import org.firstinspires.ftc.teamcode.systems.ColorSensorSystem;
+import org.firstinspires.ftc.teamcode.systems.Direction;
 import org.firstinspires.ftc.teamcode.systems.MecanumDriveSystem;
 
 
@@ -31,10 +32,10 @@ public class BlueTop extends LinearOpMode {
         this.colorSensorSystem.goDown();
         sleep(1500);
         if(colorSensorSystem.isRed()) {
-            this.mecanumDriveSystem.driveBackwards(0.5, 0.5);
+            this.mecanumDriveSystem.drive(0.5, 0.5, Direction.REVERSE);
             sleep(250);
         } else if(colorSensorSystem.isBlue()) {
-            this.mecanumDriveSystem.driveForward(0.5, 0.5);
+            this.mecanumDriveSystem.drive(0.5, 0.5, Direction.FORWARD);
             sleep(250);
         } else {
             mecanumDriveSystem.stop();
@@ -54,7 +55,7 @@ public class BlueTop extends LinearOpMode {
             knockDownRedJewel();
             mecanumDriveSystem.stop();
             sleep(1000);
-            this.mecanumDriveSystem.driveForward(0.5, 0.3);
+            this.mecanumDriveSystem.drive(0.5, 0.3, Direction.FORWARD);
             sleep(1000);
             mecanumDriveSystem.stop();
             count++;
