+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/systems/Direction.java
@@ -1,0 +1,9 @@
+package org.firstinspires.ftc.teamcode.systems;
+
+/**
+ * Created by Mahim on 1/20/2018.
+ */
+
+public enum Direction {
+    FORWARD, REVERSE
+}
